+++ CrystalScene/Command/PolygonMeshAddFacesCommand.cpp
@@ -40,7 +40,7 @@ bool PolygonMeshAddFacesCommand::execute(World* world)
 	auto shape = scene->getShape();
 	const auto vertexIds = args.vertexIds.getValue();
 
-	const auto faceId = shape->getFaces().size();
+	const auto faceId = static_cast<int>( shape->getFaces().size() );
 	Face f(vertexIds[0], vertexIds[1], vertexIds[2], faceId);
 	shape->faces.push_back(f);
